+++ Demo/Sources/Demos/Demo Screens/WidthCustomizationViewController.swift
@@ -39,7 +39,7 @@ final class WidthCustomizationViewController : UIViewController
                 
                 section += Item(
                     CardElement(title: "Default Row In Default Section", color: .white(0.95)),
-                    sizing: .thatFits
+                    sizing: .thatFits()
                 )
             }
             
@@ -49,7 +49,7 @@ final class WidthCustomizationViewController : UIViewController
                 
                 section += Item(
                     CardElement(title: "Default Row In Fill Section", color: .white(0.95)),
-                    sizing: .thatFits
+                    sizing: .thatFits()
                 )
             }
             
@@ -65,13 +65,13 @@ final class WidthCustomizationViewController : UIViewController
                 
                 section += Item(
                     CardElement(title: "Default Row In Left Section", color: .white(0.95)),
-                    sizing: .thatFits,
+                    sizing: .thatFits(),
                     layout: .init(width: .default)
                 )
                 
                 section += Item(
                     CardElement(title: "Left Aligned In Left Section", color: .white(0.95)),
-                    sizing: .thatFits,
+                    sizing: .thatFits(),
                     layout: .init(width: .custom(.init(
                         padding: HorizontalPadding(uniform: 10.0),
                         width: .atMost(200.0),
@@ -82,7 +82,7 @@ final class WidthCustomizationViewController : UIViewController
                 
                 section += Item(
                     CardElement(title: "Center Aligned In Left Section", color: .white(0.95)),
-                    sizing: .thatFits,
+                    sizing: .thatFits(),
                     layout: .init(width: .custom(.init(
                         padding: HorizontalPadding(uniform: 10.0),
                         width: .atMost(200.0),
@@ -93,7 +93,7 @@ final class WidthCustomizationViewController : UIViewController
                 
                 section += Item(
                     CardElement(title: "Right Aligned In Left Section", color: .white(0.95)),
-                    sizing: .thatFits,
+                    sizing: .thatFits(),
                     layout: .init(width: .custom(.init(
                         padding: HorizontalPadding(uniform: 10.0),
                         width: .atMost(200.0),