+++ Listable/Sources/Item/Item.swift
@@ -102,7 +102,7 @@ public struct Item<Content:ItemContent> : AnyItem
         } else if let sizing = content.defaultItemProperties.sizing {
             self.sizing = sizing
         } else {
-            self.sizing = .thatFitsWith(.init(.atLeast(.default)))
+            self.sizing = .thatFits(.init(.atLeast(.default)))
         }
         
         if let layout = layout {
@@ -357,7 +357,7 @@ public enum ItemSelectionStyle : Equatable
     case tappable
     
     /// The item is persistently selectable. Once the user lifts their finger, the item is maintained.
-    case selectable(isSelected : Bool)
+    case selectable(isSelected : Bool = false)
     
     var isSelected : Bool {
         switch self {